add helper methods for collection tab
